+++ src/components/func-component.ts
@@ -1,5 +1,5 @@
 import Message from '../engine/message';
-import Component from '../engine/ecs-component';
+import Component from '../engine/component';
 import { QueryCondition, queryConditionCheck } from '../utils/query-condition';
 
 
@@ -22,8 +22,8 @@ export class FuncComponent<T = void> extends Component<T> {
 	private onMessageConditionalHandlers = new Map<string, Set<MessageCaptureContext<T>>>();
 	private onUpdateFunc: (cmp: Component<T>, delta: number, absolute: number) => void = null;
 	private onFixedUpdateFunc: (cmp: Component<T>, delta: number, absolute: number) => void = null;
-	private onRemoveFunc: (cmp: Component<T>) => void = null;
 	private onDetachFunc: (cmp: Component<T>) => void = null;
+	private onRemoveFunc: (cmp: Component<T>) => void = null;
 	private onFinishFunc: (cmp: Component<T>) => void = null;
 
 	/**
@@ -101,18 +101,18 @@ export class FuncComponent<T = void> extends Component<T> {
 	}
 
 	/**
-	 * Registers a function that will be invoked for onRemove()
-	 */
-	doOnRemove(func: (cmp: FuncComponent<T>) => void): FuncComponent<T> {
-		this.onRemoveFunc = func;
-		return this;
-	}
-
-	/**
 	 * Registers a function that will be invoked for onDetach()
 	 */
 	doOnDetach(func: (cmp: FuncComponent<T>) => void): FuncComponent<T> {
 		this.onDetachFunc = func;
+		return this;
+	}
+
+	/**
+	 * Registers a function that will be invoked for onRemove()
+	 */
+	doOnRemove(func: (cmp: FuncComponent<T>) => void): FuncComponent<T> {
+		this.onRemoveFunc = func;
 		return this;
 	}
 
@@ -199,15 +199,15 @@ export class FuncComponent<T = void> extends Component<T> {
 		}
 	}
 
-	onRemove() {
-		if (this.onRemoveFunc != null) {
-			this.onRemoveFunc(this);
-		}
-	}
-
 	onDetach() {
 		if (this.onDetachFunc != null) {
 			this.onDetachFunc(this);
+		}
+	}
+
+	onRemove() {
+		if(this.onRemoveFunc != null) {
+			this.onRemoveFunc(this);
 		}
 	}