+++ src/index.ts
@@ -1,9 +1,10 @@
 import Engine, { GameLoopType, EngineConfig } from './engine/engine';
-import Component from './engine/ecs-component';
-import { Messages, AttributeChangeMessage, StateChangeMessage, FlagChangeMessage, TagChangeMessage } from './engine/ecs-constants';
+import Component from './engine/component';
+import { Messages, AttributeChangeMessage, StateChangeMessage, FlagChangeMessage, TagChangeMessage } from './engine/constants';
 import Flags from './utils/flags';
 import GameObjectProxy from './engine/game-object-proxy';
 import Message from './engine/message';
+
 import GameObject from './engine/game-object';
 import BitmapText from './engine/game-objects/bitmap-text';
 import Container from './engine/game-objects/container';
@@ -16,7 +17,7 @@ import Text from './engine/game-objects/text';
 import TilingSprite from './engine/game-objects/tiling-sprite';
 
 import Builder from './engine/builder';
-import Scene from './engine/ecs-scene';
+import Scene from './engine/scene';
 import AsyncComponent from './components/async-component';
 import ChainComponent from './components/chain-component';
 import DebugComponent from './components/debug-component';